flask-server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.0.5
Werkzeug==2.3.7
yarl==1.9.2
